Fix bugs and improve test suite.
