+++ .azure/azure-linux-template.yml
@@ -103,7 +103,7 @@ jobs:
           ;;
         *)
           stack ./examples/counter.hs check
-          stack test --flag reanimate:test
+          stack test --flag reanimate:test --test-arguments="-j2"
           ;;
       esac
       set +ex
