+++ .azure/azure-osx-template.yml
@@ -32,7 +32,7 @@ jobs:
       export PATH=$HOME/.local/bin:$PATH
       stack --install-ghc test $ARGS --flag reanimate:test --only-dependencies
       stack ./examples/counter.hs check
-      stack test --flag reanimate:test
+      stack test --flag reanimate:test --test-arguments="-j2"
     env:
       OS_NAME: ${{ parameters.os }}
     displayName: 'Installation ${{parameters.os}} & Test'
