+++ .azure/azure-windows-template.yml
@@ -70,8 +70,6 @@ jobs:
   - bash: |
       set -o xtrace
       stack ./examples/counter.hs check
-      stack exec -- ghc ./examples/counter.hs
-      ./examples/counter check
       rm examples/*.golden
       stack test --flag reanimate:test --test-arguments="-j2" --fast --extra-lib-dirs="$CONDA/Library/bin/"
     env:
