+++ reanimate.cabal
@@ -109,6 +109,7 @@ test-suite spec
   other-modules:
     UnitTests
     Properties
+    Helpers
   hs-source-dirs: test
   if !flag(test) {
     Buildable: False
